Refactor method move()
Removed the redundant switch statement, which
basically checks validity of move and get the new
xy-coordinate twice, both for desiredDirection and
currentDirection. Instead, I've split the logic
into smaller functions. Two for getting the new x
and y coordinates and isMoveValid() method for
checking validity only. Now, validity is checked
only once if desiredMove is already valid using an
if-statement and a boolean flag and assigns it to
currentDirection.