Change logic for timing and pacman display
Modified the logic for movement so that instead of
using delta time, which I am not knowledgeable of,
I just used the actual frames and modulo operator
as indicator of when to update pacman's location.
Also modified painting pacman's location. Added
a boolean mouthOpen to allow pacman's mouth to be
a bit animated every 10 or so frames.